requirements.txt
requests
python-telegram-bot==21.4
python-dotenv
openai